pyproject.toml
[project]
name = "crear-agentes-de-ai-con-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "huggingface-hub>=0.36.0",
    "ipython>=8.0.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.0.7",
    "langchain-anthropic>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-google-genai>=3.1.0",
    "langchain-google-vertexai>=3.0.3",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "torchaudio",
    "torchvision",
    "transformers>=4.57.1",
]

[dependency-groups]
dev = [
    "grandalf>=0.8",
    "ipykernel>=7.1.0",
    "langgraph-cli[inmem]>=0.4.7",
]
[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
